requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
anthropic>=0.40.0
pyyaml>=6.0
scipy>=1.11.0
scikit-learn>=1.3.0
